Add invalid memory access test.
- Add 2 test cases.
- Add JSON Schema for invalid memory access.
- Add assertions for runtime errors.
- Fix parsing address.
- Refactoring.

+++ include/util/address_convert.h
@@ -12,7 +12,7 @@ class address_convert
 {
 public:
     static uint16_t get_address(emulation_devices *device, json value);
-    static uint8_t to_byte(emulation_devices *device, json value);
-    static uint8_t to_two_complement_byte(emulation_devices *device, json value);
+    static uint16_t to_byte(emulation_devices *device, json value);
+    static uint16_t to_two_complement_byte(emulation_devices *device, json value);
     static string to_hex_string(uint16_t value);
 };
